# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cubit Control Plane API
  version: 1.0.0
  description: >-
    Strict REST API for Cubit, an open-source bare-metal PaaS for running
    Cloudflare Workers and Durable Objects using celld, Traefik, and S3 storage.
servers:
  - url: /api/v1
    description: Cubit API v1
paths:
...
  /r2/buckets:
    get:
      summary: List all R2 storage buckets
      operationId: listR2Buckets
      tags:
        - r2
      responses:
        '200':
          description: List of R2 storage buckets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/R2Bucket'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create a new custom R2 bucket
      operationId: createR2Bucket
      tags:
        - r2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateR2BucketRequest'
      responses:
        '201':
          description: Bucket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/R2Bucket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /r2/buckets/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Delete an R2 bucket (protected fleet buckets cannot be deleted)
      operationId: deleteR2Bucket
      tags:
        - r2
      responses:
        '204':
          description: Bucket deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /r2/buckets/{name}/objects:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List all objects in an R2 bucket
      operationId: listR2Objects
      tags:
        - r2
      responses:
        '200':
          description: List of objects in the bucket
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/R2Object'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /r2/buckets/{name}/upload:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Upload an object into an R2 bucket
      operationId: uploadR2Object
      tags:
        - r2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadR2ObjectRequest'
      responses:
        '200':
          description: Object uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  key:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /r2/buckets/{name}/objects/{key}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get raw object contents from an R2 bucket
      operationId: getR2Object
      tags:
        - r2
      responses:
        '200':
          description: Object binary or text content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete an object from an R2 bucket
      operationId: deleteR2Object
      tags:
        - r2
      responses:
        '204':
          description: Object deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  responses:
    BadRequest:
      description: Bad request / validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with existing resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden operation on protected or restricted resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    R2Bucket:
      type: object
      required:
        - name
        - createdAt
        - objectsCount
        - sizeBytes
      properties:
        name:
          type: string
          example: "user-assets"
        objectsCount:
          type: integer
          example: 4
        sizeBytes:
          type: integer
          format: int64
          example: 1048576
        isSystem:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
    CreateR2BucketRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "user-assets"
    R2Object:
      type: object
      required:
        - key
        - sizeBytes
        - contentType
        - etag
        - lastModified
      properties:
        key:
          type: string
          example: "avatars/user-123.png"
        sizeBytes:
          type: integer
          format: int64
          example: 20480
        contentType:
          type: string
          example: "image/png"
        etag:
          type: string
          example: "\"d41d8cd98f00b204e9800998ecf8427e\""
        lastModified:
          type: string
          format: date-time
    UploadR2ObjectRequest:
      type: object
      required:
        - key
        - content
      properties:
        key:
          type: string
          example: "data.json"
        content:
          type: string
          description: Raw or base64 text content of the object
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "Application with ID not found"
        details:
          type: array
          items:
            type: string
          nullable: true